feat: Created YearsController and Service, moved year-based methods to its own API endpoint, added getEntities() and changed getByYear to getById in 'years' endpoint. Added movies list in years table with relation oneToMany mappedBy

+++ src/main/java/dev/daniel/movies/dtos/MovieDTOResponse.java
@@ -1,4 +1,4 @@
 package dev.daniel.movies.dtos;
 
-public record MovieDTOResponse(Long id, String title, int release_year) {
+public record MovieDTOResponse(Long id, String title, Long release_year) {
 }
