+++ src/main/java/dev/daniel/years/YearsRepository.java
@@ -2,6 +2,6 @@ package dev.daniel.years;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface YearsRepository extends JpaRepository<ReleaseYearEntity, Integer> {
+public interface YearsRepository extends JpaRepository<ReleaseYearEntity, Long> {
 
 }
